Decouple reading EEPROM from writing EEPROM.  Now writeEEPROM() does only what it's name implies and it's method takes no arguments.  This avoids the 0/1 magic numbers which was a flag to decide whether the user should be notified after the EEPROM has been read (!).
Introduced a method called readEEPROMAndNotify() which is used in all
cases where writeEEPROM(1) was previously called.

Additionally this avoid re-reading the profile when the reboot command
is issued which speeds up reboots.

Finally this avoid needless comments because the code tells you now what
it is going to do.

Conflicts:

	src/cli.c
	src/config.c
	src/config.h
	src/mw.c
	src/sensors_acceleration.c
	src/sensors_compass.c
	src/serial.c