Don't restrict current config
ACS756 current sensor has an offset of 2.5V and it's output is either -40mV/A or 40mV/A. If you wire it that the currentscale is negative you should never get values >2.5V out of it. It gives 2.5V at 0A and 0V at 62.5A. To be able to use it like that you need currentoffset=2500 and currentscale=-400.